wx: allow subsetting in GUI
WIP: there are difficulties with using boost date times, and deciphering
when and where times are local and when they are in UTC.  Currently we
display times in the GUI as UTC, and also use them in UTC.

We should probably use a more complex model in the interface to show
local times, but store the boost time object in the background.  Right
now we only store the string representation.

TODO:

- cli support (allow both local and utc input)
- proper zone handling in display
- documentation