nomads: add hrrr subhourly support
Intial implementation to download/add sub hourly (HRRR-SR).  The HRRR-SR
does not have the TCDC variable, but it may have proxies for cloud
cover.  All cloud cover layers are set to 0% coverage for all runs.  If
and when we figure out a decent proxy, it may be possible to add.

The previous initialization was strict about using the n+1 timestep in
GFS to set cloud cover data for the 0th time step.  The code is now less
strict, allowing the default setting of 0% cloud cover.

Fixes #240