gui: initial port to Qt5
This ports the existing gui to Qt5 without creating the interface from
scratch.  The build system has been updated, and the include directives
for the GUI files have been modified as needed.

The splash screen no longer works as intended, due to a Qt API change.
This will be fixed later.

On a crappier note, building on Ubuntu 20.04 causes what appears to be
an invalid write or memory corruption while setting some location data.
This manifests itself by setting a latitude value out of the range -90
to 90.

This causes an exception to be thrown and the run aborts in the command
line version.  In the GUI the exception is not caught.

The real story is that we have to find why the latitude value is set to
the longitude value, one valgrind run failed to turn it up.