ninja: disable experimental gdal_output by default
CentOS 7 still uses GDAL 1.11.x, from ~5 years ago.  That distro is used
for many a HPC system.  The gdal_output code depends on the next
release, GDAL 2.x and higher.

The option to build gdal_output is added to cmake, and defaults to OFF.
It is currently not used in the actual codebase anyway.