Updated the /src/ninjafoam/ libWindNinja.so style libraries and applyInit applications to be the ones that I used for OpenFOAM 8 simulations in my porous media project for simulations without porous media. In this process, I deleted any old libraries that were not being used anymore, I tried to only update the ones that were currently being used by copying and pasting the code from my OpenFOAM 8 versions into the existing files replacing all the text, deleting all the unused libraries.
What is done in the code, the formulas, the basic lines of code, are mostly the same, in most cases identical to the old code. But when I first converted them to OpenFOAM 8 during the earlier days of my porous media project, I did a little reorganizing and formatting to make them easier to read. I would have preferred to do this transferring process little by little, just updating code as the snippets changed instead of copying and pasting what I had, preserving how the code is different in git, but the formatting was so different and some of the reorganization was so different, that it just wasn't worth it. So you can still compare this code to the old code, but you'll want to look at the code itself for the two different versions instead of comparing using the git diff (which is what is seen when looking at the commit on git).

I didn't add my updated epsilonWallFunction or my updated kEpsilon model, called myEpsilonWallFunction and myWindNinjakEpsilon, because when I first tested the conversion of WindNinja stuff to OpenFOAM 8 I made all those changes but didn't see that much difference in the results, I think it would be good to test how well the internal OpenFOAM 8 updated kEpsilon model and epsilonWallFunction libraries do before trying my project updated versions.

I did add a new wallBC, nutkAtmRoughWallFunction, called myNutkAtmRoughWallFunction in my original porous media libraries, to the /src/ninjafoam/ folder of libWindNinja.so style libraries. The inheritance structure for the nut wall functions changed quite a bit when going from OpenFOAM 2.2.0 to OpenFOAM 8, and while nutkAtmRoughWallFunction existed as a wall function in OpenFOAM 2.2.0, it stopped existing, or at least became an updated version too integrated with other things, in OpenFOAM 8. I think it is now the /opt/openfoam8/src/atmosphericModels/derivedFvPatchFields/nutkAtmRoughWallFunction/, but it appears designed to only work when using all the other atmBoundaryLayer functions found in that folder, if I remember correctly they didn't seem built to allow using a single function from there.
Anyhow, if I remember correctly, I checked all the inheritance layers in both versions of OpenFOAM to make sure that my new nutkAtmRoughWallFunction would work correctly, using the same formulas as the old one, and the resulting values seemed to be similar. I believe using the nut wall function from the OpenFOAM 8 windAroundBuildings tutorial (nutkWallFunction) behaved more cyclical, but overall didn't seem to change the results that much, but it definitely was not exactly the same. For now let's use my version and go from there if there are problems.
Might be good to rename this wall BC, which would require updating code within ninjaFoam.cpp to place the right entry for differing versions of OpenFOAM, to avoid conflicts with the libatmosphericModels.so containing the internal OpenFOAM 8 version of nutkAtmRoughWallFunction, but this seems to only be a problem if including both libWindNinja.so and libatmosphericModels.so at the same time within the same simulation which doesn't seem likely to occur.

That wraps up everything that I have ready made from my porous media project other than maybe using some of the /system/fvSchemes and /system/fvSolution stuff. Also using my insert building via snappyHexMesh files as examples to figure out how to update moveDynamicMeshDict, topoSetDict, and refineMeshDict. I also have some form of sample files that have different names, if they are the same thing then the file name changed from sampleDict to singleGraph, I'm not sure if they are related though. In the past when trying to do the conversions, I tended to look at tutorials from both versions and compare files that seemed similar if a file no longer seemed to exist till I found what worked. So there is still a lot to do, but the files that could be transferred over that were already ready to go have all been added.

For Issue #458 [ci skip]