Attempt to add build dependency cacheing to the github actions
If it does what I expect/hope it will do, it should run build_deps during this commit, saving the poppler, proj, and gdal dependency folders with their contents. During the next commit github actions should restore those existing cached folders instead of running build_deps, meaning the testing will fail because there are many other dependencies built by build_deps that are not cached (apt install stuff, qt, openfoam).

In practice, I know that the | symbol lets you do multi commands in a single shell, so it is supposed to be cacheing three separate directories, and while I think that I've got that part right, I am a bit leery about how the key is setup, I think the hashFiles() thing is confusing, the tutorials made it sound like maybe it used that part to know when it needed recached or not hence why it wanted more than just a single string value but I'm not sure.

There may be other things wrong with the attempt as well. There was something about caches staying in the memory until forcefully deleted, so a periodic cleanup of cacheing may be required, or a cleanup each time a change is done to the build dependencies.