created new output class similar to volVtk, for outputting a wrf-like structured netcdf file output for the mass only solver.
this class will likely need adjusted as the desired requirements for output netcdf are refined. Also, work to do the same type of output, and an unstructured output ugrid, are currently being worked on for the OpenFOAM solver.

The output netcdf file is currently hard coded, it wouldn't be too much work to add inputs for it, but as the netcdf format is still a work in progress, it makes it easier to just have the hard coded output for now.