Figured out the problem with the lat/long conversion for the mass solver netcdf output converting from utm to lat/long points. Turns out to have been that the points needed converted from local (0,0) as the bottom left corner to the original projection (utm in this case) corner value as the bottom left corner value.
This correction has the netcdf output points now working as the correct lat/long, turns out that the long points were a bit off before as well. And the conversion works with each of the various ways to set the spatial reference for the transform, either using projection string or to get the spatial reference from EPSG, I've set it to what seems to be the best while leaving the remaining commentable code as examples of other ways to set the spatial reference for the transform.

Added a CPLFree() for the WKT string used for getting the lat/long projection string, turns out the gdal documentation requires it.