Trying yet again to fix github actions
The simple fix didn't work, says it is already the latest version of the odbc libraries. So I'm moving on to try the more complex fix, which also likely will not work because it's just trying to tweak some of the underlying environment packages without fully messing with the ones built with the past packages.

For sure, there are now NEW errors, somehow apt install was not able to find libcurl4-gnutls-dev, and something seemed to also go wrong with boost? gdal had trouble building in the same way as before, but WindNinja cmake broke with a boost error instead of a gdal error as before. Something about boost_unit_test_framework-config.cmake not being found within the boost_unit_test_framework, though boost seemed to at least build like normal.

Anyhow, this fix should at least hopefully fix the gdal build problem, if it works. Can go from there to figure out the other new boost and curl problems.