Another try at fixing github actions
Looking over the build logs again, I found that when the libcurl4-gnutls-dev failed to install, the full list of apt-get install commands was abruptly stopping in the middle without installing any of the other packages. Reading online, I found another possible fix to try, which is to try running sudo apt-get update and sudo apt-get upgrade before installing any packages. The build log did show 14 packages in the background not yet upgraded, so I thought it would be worth a try.