Test of fixed github actions
Looks like the sudo apt update and upgrade was the missing key. Now I'm going to test two things. 1) seeing if the odbc stuff was needed after all, so dropping that, as the output from the last run seemed to imply that the command to update odbc was ignored, and I guess the problem ended up somehow getting fixed by sudo apt update and upgrade instead. 2) I want to see if it uses the cached files or tries to build a new cache, I'm expecting it to build from the cache.