Test github actions cacheing
Moved the cache from /windninja/ to the /windninja/scripts/ directory. The test is that as the files/folders to cache have changed, this SHOULD make it rebuild the cache. Not sure what happens to the old cache files though, like is it smart enough to realize to delete them now that they have a new folder name even though the old folder names are no longer given. I still haven't tracked down how to look at the caches similar to looking at the runner image, currently I just see whether files get cached and whether they are grabbed and the paths on the runner image they end up using.