removed all non-standard huvw output options since there are already enough client-specific ones. All we need is to store the huvw GDAL dataset.
Added a huvw0 option, which stores the mesh init from wxModels or stations. Useful to see diffs

added a 'spd' raster band to huvw data sets. It is redundant but allows to create contour maps with gdal_contour

gdal_util.* still have to be cleaned up since a number of functions are not used anymore now that huvw post-processing is external