a fix for a mistake made in the if NULL check that I did to try to improve the pointInitialization::FetchStation() bad station error messages for if the GDAL getFeature() function breaks in that same way again. For issue #506
the problem was with the consistency of vector sizes for the vectors defined and pushed to at various stages of the function, by adding a new if statement adding to the stationChecks vector without adding some kind of default bad station value to other similar vectors, these vectors would become mismatched in size. Looks like the only vectors in danger of this at the scope I added the if NULL check to, were mnetid and stationCSVNames

Turns out that stationCSVNames was purposefully kept a different size and only added to in the moment a station was finally to be written to file (so excluding bad stations). mnetid could have been mismatched at calls to the if mnetid[ex] == 1 or 2, so I added a default value of 0 to that vector so that any accidental calls to if mnetid[ex] == 1 or 2 to be skipped