for momentum solver mass mesh vtk output
found the problem with using finer mass mesh resolutions, turned out that input.dem and input.surface are resampled to the mass mesh resolution during massMesh.buildStandardMesh(), but the init->L and init->bl_height were not resampled, so referencing the L and bl_height arrays with the higher resolution led to referencing outside the arrays. Fixed it by resampling the additional arrays

tested the fix for two different dems, with coarse and fine ninjafoam simulations, with various mass mesh resolutions set using both methods. In all cases, no additional changes to the method, especially the no data filling method, were required. Maybe if trying to make meshes at resolutions significantly larger than the ninjafoam case res, but so far it hasn't popped up in any of the cases I tried

still one major difficulty, seems like the mass mesh is NOT the same resolution, and therefore points, as a mesh generated by the mass solver. This seems to be because of resampling the dem and other ascii grids multiple times, in different ways for the two solvers, ninjafoam in particular seems to resample it at least one additional time than the mass solver before reaching writeOutputFiles(). Not sure of which will be the cleanest fix for that yet.

found two additional small one line fixes to the method, small typos