for momentum solver mass mesh vtk output
still needs to be tested for other resolutions and cases, but seem to have come up with a fix for the differing mesh resolutions and x,y,z points of the massMesh vs a mass solver run mesh. Trick was to construct a mass mesh from a dem itself (careful with extra dependencies), and to use a copy of the starting dem and other ascii grids for the resample to final massMesh resolutions. The issue seemed to be that for mass solver runs, the ascii grids are resampled from dem resolution to mass solver run resolution, then to whatever is needed for output, but for momentum solver runs the ascii grids are resampled from dem resolution to momentum solver run resolution then to whatever is needed for output, always resampling from dem resolution to mass solver run resolution then to output seems to have fixed the differences in the mesh resolutions and x,y,z points

That being said, I ran into trouble making a copy of the init->L and init->bl_height ascii grids, the boost::shared_ptr init caused me problems with the various copy constructors and I couldn't figure out what I was doing wrong. Luckily, those two grids are not resampled in the same way as the other grids, so I think that skipping the copy of those two grids and resampling them as is, is still a good solution, at least for now