for momentum solver mass mesh vtk output
changing method from working for specified mesh resolution, to always expecting the momentum solver mesh resolution to be custom resolution set to be that of the mass solver mesh resolution

also setup the foam mass mesh vtk output flag to now be a CPL config option to get it running rather than having a hard coded bool. Also did a hint of cleanup

when debugging this, turned out that while the resulting x and y points looked right, the domain height was off. To fix this, the domain height has to be computed from the raw dem BEFORE calls to resample the dem. The fix started to smell like a copy of the ascii grid, just computing and storing the domain height earlier, but do avoid such copies I just moved the definition of the massMesh to right before the dem gets resampled and right after the calculation of the mesh resolution

in this process, I was also able to drop all the resampling of the various ascii grids, as now they are all resampled once, right to the input custom set mesh resolution, the L and bl_height grids specifically being defined at the mesh resolution at construction time. There was still a risk that the resampling of the dem and input.surface ascii grids would be done twice, but I checked and for the coarse and fine cases that I checked, the second resampling is not called as the grids are already at the mesh resolution