fixed unit conversion problem with the turbulence col max grid output
turned out it was actually converting fine, was actually an AsciiGrid::ascii2png() problem. The method for calculating a scaled grid of int values between 0 and 255 from the min and max double values was off, making the scaled grid go to values above 255, making the plot drop out those values. Turning the temporary 0.6 scale value to 0.4 helped fix this, but saw another issue with the min and max value sometimes going outside the range of the int values from rounding, dropping a very small subset of values for some cases. In the end, I found it easier to just replace that whole method with a histogram style binning of values

caught another small issue, I hope my fix was good enough for it. The int scaled grid of value breaks were NOT matching the legend val grid breaks, was 0.2 for one while being 0.25 for the other. I found that changing the scaled grid to match the color legend made the plot change significantly, while changing the color legend to match the scaled grid just made slight changes to the legend labels, so I went with the latter as my fix. Note that the chosen breaks of 0, 0.2, (0.2+1)/2 = 3/5, 1 as well the older color legend breaks of 0, 0.25, (0.25+1)/2 = 5/8, 1 seem to be kind of strange choices to me, not sure the reasoning behind using these breaks as opposed to other breaks so for now I'm keeping them as they are. That being said, I believe the 2 colors between the main colors for each break in the bmp legend colors does still make sense in that the colors vary linearly within a set of breaks, but I still think it might still be a hint off in how the user might interpret the scale since it isn't necessarily linear across breaks. The fact that it IS linear within a set of breaks probably makes it good enough for now though