adjusted new ascii2png command line color ramp setting methods to be better about swapping back and forth between the methods, and allowing default values. Also adjusted the default values a bit. Adjusted ascii2png() to allow minToMax_uniform method as well.
still not the most happy with the command line stuff, I rewrote it three more times just to get to this point though. I think that wrapping the colorRamp stuff into a class would simplify some of the default setting in the input classes, but this definitely works good enough if we aren't planning on changing it more for a while

finally wrapped this up to a finalized enough state. Phew