found and did a partial fix to the bounds for the kml png layers
the original was only using the transform of the southwest and northeast corners of the ascii grid, which are no longer the min and max corners of the ascii dataset after transformation. The fix is only partial because it will break when longitude crosses between -180 and 180 degrees, because the min and max methods swap in weird ways there.