added standalone hrrr_to_kmz script, for reading in hrrr grib2 files and outputting the corresponding kmz files, without having to run WindNinja itself each and every time
this script differs from the wrf_to_kmz script, in that the dataset starts out as the entire conus region instead of already being downsized to a region of interest. So this script has an additional step to clip the data to a smaller region of interest

there are still a few things to work out. Currently the script uses hard coded lat/lon points for the bounding box, the script still needs the inputs setup for various clipping box methods. In addition, the clipping is currently done using ascii grid clipping, after warping the dataset to lat/lon, but the clipping should probably be done using gdal itself in earlier steps