cleaned up the inputs processing for the hrrr_to_kmz script a bit more. Switch from cout to printf to get a few more decimals displayed for some of the print statements. Added checks for the input lat/lon points.
That being said, there seem to be two major checks to the inputs left that could be done. 1) open the dataset and check the clip box fits within the dataset bounds. 2) all the lat/lon stuff, including the ascii grid clipping, need revamped to handle bounding boxes that cross the -180 to 180 longitudes, not often that the area of interest bounds those longitudes but it is important to be aware that my lat/lon methods currently don't handle longitude crossings.