slight modification to past fix for 0 input wind speed momentum solver cases, the if( input.inputSpeed != 0 ) worked, but doing a check against 0.0 is more correct.
did notice, that while this ninjafoam method to skip simpleFoam works, the outcome is still off in that the turbulence output uses the k sampled from the starting time directory, and so while the wind speed comes out to be zero or the diurnal values from the mass solver, the sampled turbulence comes out to be a high set of values because the initial k value is 50 m^2/s^2 not zero.

intent is to shift to skipping ninjafoam simulateWinds() completely using ninjaArmy, but wanted to put a record of this fix here just in case. For issue #526