Update build_deps_ubuntu_2204.sh
cleaned up a bit more of the comments and whitespace, and a few of the commands including the apt-get vs apt stuff. Just need to do one last test of the script on a fresh VM of Ubuntu 22.04, then, unless there are other proposed edits, ready to merge with the qt4guiForUbuntu22o04 branch, replacing the script that I had built using the old poppler and proj dependencies.

Still have the qt4 problems to work out, but from what I can tell everything else is working with the changes that we've made so far.