added pointInitialization wxStation datum check, which wasn't happening before. In the process, found the way pointInitialization error checking was handled in the gui needed updated to catch specific error messages for deeper layers like wxStation.cpp function calls.
Note that the wxStation::check_station() datumType check failed to run correctly, because the datumType is an enum set to one of the expected values at initialization, it can never have an unexpected value. I did try setting datumType to NULL at constructor time, to cause the wxStation::check_station() datumType check to finally run, but this just resulted in datumType taking the value of the first item in the enum list, so that check still failed to run.

The in the moment check I added to wxStation::set_location_LatLong() that uses and sets the datumType is more useful anyways, as it disallows calls to GDALPointFromLatLon() without a proper datum value.