fixed gui throw error/warning for when solve is somehow run with a blank output dir
this is NOT the fix for issue #540 yet, but is related to it. Qt was disallowing throw, so I updated the method to be the qt style. So now running solve() with an empty output dir doesn't die with a seg fault, but warns and just stops the solve process.