stopped gui output path tool from setting the output path to a blank string when users click cancel, now clicking cancel always sets the output path to the starting path instead. For issue #540.
also fixed a bug where the start path for the gui output path tool was always one directory up from the starting path.

note that the start path for the gui output path when WindNinja is first started up, is always an empty string, but solve cannot be selected until the user specifies a dem. Each time the user selects a new dem, the output path always gets set to be the path of the new dem. Related to issue #369