fixed srtm downloads to no longer need to fill no data values, by requesting a buffered bbox, warping, then clipping to a final warped bbox, rather than the original method of requesting the original bbox, warping, then nan filling.
To do the buffering, I used the wxModelInitialization::ComputeWxModelBuffer(), Lx*0.2, Ly*0.2 as the buffer. Tested with what should have been almost the maximum 100 km x 100 km size limit, and this buffer method seems to be holding up.

For the warping and clipping, I ended up using the method from GDALFetch::FetchBoundingBox(), almost a straight copy and paste of that method, with slight cleanup of some of the warning messages. I originally tried out a new method, with gdal_utils.h GDALWarp(), which involved the strange GDALWarpAppOptionsNew() as well as an intermediate shapefile for the bbox to clip to, but it ran a heck of a lot slower for larger datasets, and outputting the bbox to a shapefile then rereading it in seems kind of roundabout compared to the GDALFetch::FetchBoundingBox() method currently implemented.

for issue #529