foam breaking cases fix attempt, an ability to do a small nudge to the cell height
might be because of the slight nudge to the color scale with a slightly differing result, but running with 0.9 and 1.1 times the blockMesh height led to the resulting pdf for bell_steep.tif, run as fine, brush, 5 mph at 270 degrees, to have differences in the solution. In particular, the wake stopped being symmetrical. Interestingly, the cell height near the ground seemed to only experience a slight nudge in height change, the upper cells slightly more so, the overall combined nudges in cell height making about a cell height in change to the overall blockMesh height, so it really wasn't that much of a change to the mesh at all for such surprisingly differing results.

So I'm kind of leery of this solution attempt for breaking cases. Though maybe it's just that something is going on with the log profile interpolation sampling? Well, it's an attempt at least.