fix for the newly added foam cell height nudging method, the problem described in bee19004205778a5bd1e895d7e70380a48e9948b
turned out that the problem was that while the cell heights were being modified, the initialFirstCellHeight was not being modified alongside the changes to the mesh. This resulted in refineMesh not splitting cells in the same way, some cell regions especially in the taller terrain regions were not being split.

That being said, turns out that the original run with no changes to the mesh, for the bell_steep.tif case, was ALSO not splitting cells properly, the top of the hill and the edges just off of the top of the hill are NOT currently getting split with our current foam implementation. So the initialFirstCellHeight turns out to be a poor estimate even for regular fine runs for that .tif file. But, without this fix, even more regions were not getting split correctly, with this fix it now appears to be at least replicating the regular case cell splitting.

I do remember trying out a method to completely get rid of our firstCellHeight estimation problem, refineMesh using a minZpatch.stl file instead of the dem, but the method seemed to get really slow for finer meshes, seemed to cause other differences in the solution as well.