fixed error in angleFromNorth adjustment for output kmz. For issue #544
looks like kmz was the only output that needed such an adjustment. The raw values printed to the kmz needed adjusted back by angleFromNorth, the vector drawing was still good.

though it should be noted, that this does mean a slight difference of 2 degrees in drawn vectors compared to the original, as we weren't accounting for angleFromNorth for all our past utm cases. A second check to show that this ~2 degree angleFromNorth of utm vs lat/lon adjustment is correct, is to plot the vectors in lat/lon using the lat/lon angle, those plotted vectors come out the same as our current plotted vectors in direction and angle, where our current method is to plot the vectors in utm using the utm angle and projecting the vectors to lat/lon.