2nd cleanup attempt of casefile stuff continued, for issue #7
rearranged some of the inputs to casefile in mainWindow.cpp and cli.cpp, including the way ofstreams were being opened and used. Needed ofstreams to stay in scope only opening when required, instead of always opening. Similar updates to the locations in the code for file deletions, only open files when needed and if deletion is required, delete right after closing. Cleanup of the error code in mainWindow.cpp as well, have to use qt forms instead of throw in the gui

first attempt at getting the proper full list of wxModel times, when the user selects 0 times and so the code runs using all wxModel times. Still some todo/issues with this method

dropped some unused variables, functions, and dependencies in casefile as well