changes required to compile the casefile-updated branch on windows. For issue #7
while it compiled, and ran in what ways it could, all the cpl_zip() type commands had to be commented out, so it isn\'t doing any zipping stuff anymore. The error is stuff like, "unresolved external symbol to cpl_zipOpen". So it seems like a change that needs to be made to CMakeLists.txt to include zlib. I see zlib inside the included GIS_INTERNALS, but it doesn\'t seem to be getting grabbed correctly. I tried building zlib from scratch, GIS_INTERNALS is using zlib 1.2.3 but I could only get ahold of the current zlib which was version 1.3.1. I\'m not sure if the problem was this zlib version difference, or if something else was missing, maybe some edits to how gdal is included to point to zlib or another path to something more specific in zlib, but including zlib this way didn\'t seem to fix the problem. Note that in casefile, the include is "#include "cpl_minizip_zip.h"", so maybe it\'s some other slightly missed include

I think most of the other changes are good, except that the static_casts to get std::to_string(intVal) need to be checked, on both Ubuntu and Windows, to make sure that they don't mess up the printed values by adding decimals and stuff. Not sure what to do about "#include <mutex>", it wasn't Windows compatible.

Seemed like it wasn\'t outputting the intermediate log files, might be a good thing, might not, need to check and make sure that the failing of the zip file to open is properly being handled in the later intermediate file outputs.