attempted to fix the minizip "unresolved external symbol to cpl_zipOpen" type problems for windows builds of the casefile. For issue #7
I was not successful at getting minizip to work on windows, the following changes didn't hurt Ubuntu builds they just showed more specifically the libs found behind the scenes for minizip, but it was required to even attempt to get minizip through zlib to work on windows. I specified the libraries to the GIS_INTERNALS locations, then later to a hand built older minizip (https://www.winimage.com/zLibDll/minizip.html to find https://github.com/zlib-ng/minizip-ng/tree/1.2) and a handbuilt zlib of version 1.2.8, neither method fixed the problem. Seems to imply that the issue isn't minizip in zlib by itself, but that minizip in zlib has to be built correctly into gdal itself, which doesn't seem to be done for our GIS_INTERNALS windows build.

Using "dumpbin" on the ".lib" files in GIS_INTERNALS, I was able to find that there are minizip bindings somewhere in there, but apparently GDAL somehow isn't built with them, the linking is somehow broken. Because of this, I think that we are going to need to try some other zip method that is more compatible with our windows builds.