reverted casefile minizip back to standard GDAL VSI style zipping methods, first pass with GDAL zipping methods. for issue #7
still had to keep #include <mutex>, windows issue causing lib, to get this style of zipping to work without seg faulting for the 5 wxModel data times, none selected, case on Ubuntu. Still haven't found a workaround for <mutex> with the Windows build

the windows build is now compiling with zipping, but VSIRename\(\) is causing problems with the current setup. In particular, it doesn't seem to matter how the file paths are fed into VSIRename\(\), VSIRename\(\) will rename the zip file regardless of whether it has uniform or a mix of "/" or "\" style chars. But while VSIRename\(\) works on an open zip file on Ubuntu, VSIRename\(\) will fail on Windows unless the zip file is closed and reopened during the renaming process. Worse, while all the folders and files are added to the zip properly when VSIRename\(\) fails to rename the zip, when VSIRename\(\) succeeds in renaming the zip it somehow drops most of the files from within the zip.

there are pathing problems that need to be addressed for the windows build, apparently the gui grabs the dem file with "C:/Location" "/" style chars, but CPLFormFilename\(\) puts the remaining parts of the path together with "\" style chars, mixing the path char type. In addition, the paths seen in the various files within the zip get to be "\" in style, need to make sure that style makes sense.

My current method attempted to avoid reopening and reclosing the zip over and over at each addFileToZip\(\) instance, by just opening the zip once at the beginning, and closing the zip once at the end. It seemed to work, at least till I hit the windows issues, but I'm not sure whether it is wise to leave the zip file hanging open like that the whole time. It does look like the method will work though, once we get the zip file renaming and the <mutex> syncing type issues figured out. Will probably need to watch out for properly closing the zip for each instance of the program quitting early though.