figured out the windows replacement to the <mutex> library use in casefile, OpenMP omp_lock_t, which turned out to already be present all over in the WindNinja code. For issue #7
I think that this was required and works, because various ninjas were running in parallel, each trying to access and add file data to the same casefile zip file all at the same time

with this fix, as well as the last one, I tested the casefile methods like crazy on both Windows and Ubuntu. The casefile methods now work quite well in all the different kinds of ways, on both windows and Ubuntu.