updated ninjafoam sampleCloud() function, for sampling with minZpatch.stl
the method still has issues. Unlike sampleCloud() for sampling with a dem where the data are at stl points, which are at dem cell centers, the geoTransform/inverseGeoTransform being fitted to being at the dem points rather than the stl pointsl; sampleCloud() for sampling with minZpatch.stl effectively has the data at dem points, the same geoTransform/inverseGeoTransform. This results in a set of NO_DATA values at the bottom row and the right-most column, which I dealt with by doing nan filling. Ideally, the method should be adjusted to more directly place the values into all the correct positions, something is off with it. Related to this, but fixing the precision issues didn't improve this problem, I also noticed that the minZpatch.stl created .stl files have some precision issues going on, which is seen in the raw data sampled this way, so annoying.

I also noticed that the nan check in sampleRawOutput() was placed a bit late in the code, letting data with NO_DATA go into the log profile calculation results in crazy high values replacing the NO_DATA values. I moved up the nan check to a better spot in the code.