more for issue #441, replaced KmlVector::calcSplitValsFromSplitVals() lever rule methods with simpler methods. For example, used the standard calculation method for  equal_interval, and used the original averages between the given set of interval value levels for equal_count.
The resulting equal_count legend scale still doesn't match the true scale values found when just packing all the various ascii grids into one single ascii grid using the standard calculation, but it IS much closer than the lever rule method I had before. The values are still slightly larger compared to the true values, but the trend follows the true values much better than the lever rule style.

I spent a heck of a lot of time on this, there may still be a method that gets the exact scale values without packing the ascii grids together, but I couldn't find it in this short time period. I suspect that the current averaging method is as close as can be achieved with the current data being passed around, I suspect that the current averaging method is indirectly like trying to construct a very coarse histogram from averages at each given level of very few super coarse histograms constructed from only 5 values each. So the differences between the result and the true values has to do with the coarseness of the histograms, but it still follows the general trend. If the method were to be improved, probably would need to connect the index locations with the values somehow with some kind of count-like information, I couldn't quite figure it out.