first pass adding from prj1 to prj2 style calculations to each of the separate 2D surface wx model initialization classes, at the same time tossing out the incorrect adding/subtracting angleFromNorth spots in the code that I had added before back when I had been incorrectly assuming outputs and inputs were geographic rather than projected coordinates. For issue #544
to do this, I looked over and inspected each and every spot in the code that I know about up to now, including the utility scripts, to think about whether a dataset was in/supposed to be in projected or geographic coordinates, making the decision and adjusting the code at every step accordingly. Was kind of exhausting to keep track of it all tbh.

that being said, I still need to check that the from prj1 to prj2 style calculation still works, and adjust these final spots in the code accordingly, might need to use GDALDataset handles rather than pszWkt for the prj2 information. I also have NOT messed with any of the 3D initialization code yet up to this point.