code for testing the prj1 to prj2 style calculations, to figure out the final proper convention, left in the state that came out with the proper results. For issue #544
turns out to be that the coordinateTransformationAngle needs to always be calculated in either dem space, or FROM geographic space TO the given space for a given dataset. So some testing of this convention is still needed for wxModel runs where there is a third spatial reference involved, but it appears to be correct for domain average runs now. In fact, I checked this convention for momentum solver domainAverage runs of 270, 0, 90, 180 degrees on flat_test.tif (utm12N), flat_test.tif warped to utm18N, flat_test.tif warped to utm10N, and flat_test.tif warped to ConusAlbers projection. I also verified that all output products were correct with this convention for momentum solver domainAverage runs of 270 and 0 degrees flat_test.tif (utm12N), flat_test.tif warped to utm18N, big_butte_small.tif (utm12N), and big_butte_small.tif warped to utm18N.

That being said, I ran into a NEW problem with the conventions, when trying out flat_test.tif at 270 degrees warped to the Robinson World EPSG 54030 projection. The case looked very similar to flat_test.tif warped to utm18N, even the values at each step came out similar and in agreement with how the dataset appeared to be rotated. But the resulting kmz had vectors pointing the wrong way, implying that the sign of the prj2 = prj1 - coordinateTransformAngle_FROM_prj1_TO_prj2 calculation is wrong, or the sign of the coordinateTransformAngle is wrong. I tried printing the geoTransform to see if there was additional rotation that we don't account for yet, didn't see any rotation. So there must still be something off with the conventions that is NEW. For now, I'm moving on to get the conventions tested for all solver types with the projections that we know work, then we will go back and try to see what we are missing for projections like the Robinson World EPSG 54030 projection.