cleaning up the final prj1 to prj2 convention stuff from commit 7ca44ab9c748dbe99b26da1b4db945c28896413d, and propogating it to the rest of the code. For issue #544
I checked this convention for momentum solver with diurnal and mass solver no diurnal domainAverage runs of 270 and 180 degrees, on flat_test.tif (utm12N), flat_test.tif warped to utm18N, and on big_butte_small.tif (utm12N), and big_butte_small.tif warped to utm18N.

tested this convention for pointInitialization on big_butte_small.tif (utm12N), and big_butte_small.tif warped to utm18N, downloading with a 30 m radius resulted in 4 sensors outside the bounds of the dataset, I also made a test case where I placed 4 sensors within the bounds of the dataset with carefully selected values. In each of the cases the results matched closely despite using these two different projections for the dem.

tested this convention for wxModelInitialization, comparing resulting wxModel kmz outputs and simulation kmz outputs for big_butte_small.tif (utm12N), and big_butte_small.tif warped to utm18N, on flat_test.tif (utm12N), and flat_test.tif warped to utm18N, the resulting kmz outputs finally look correct! In each case, the resulting kmz vectors come out consistently shifted counter-clockwise (smaller values) by the angleFromNorth of the weather model data for the given dataset location, for both the utm12N and utm18N projected dems. For big_butte_small.tif this difference comes out to be -9.687698, for flat_test.tif this difference comes out to be -6.910659. I also tried downloading the weather model data using the utm18N projected dem, with no change in the results.

That being said, this wxModel testing still doesn't fully sort out whether the convention is completely correct or not, it doesn't explain whether the coordinateTransformationAngle needs to always be calculated in either dem space, or FROM geographic space TO the given space for a given dataset. In each of these cases the angleFromNorth values match up quite closely to the coordinateTransformationAngle values calculated in the given direction of FROM weather model space TO dem space. And for cases without a third spatial reference, like domainAverageInitialization and pointInitialization, and maybe griddedInitialization, calculating the coordinateTransformationAngle in dem space is the same thing as calculating the coordinateTransformationAngle FROM geographic space TO the given space for a given dataset. I did try pre-warping the weather model data FROM conus albers TO utm18N, but something went wrong, the wxModel dataset went off to the other side of the map, I need to figure out a better test case for this.

tested this convention for griddedInitialization, looks like it still requires some work, it requires an additional warp step if the input data is in a different projection than the dem.

tested output_converter, looks greatgit log

tested hrrr_to_kmz and wrf_to_kmz, looks like it still requires some work, but I believe that they are working as expected, as best as can be for now. hrrr_to_kmz seems to be getting a coordinateTransformationAngle of ~0.0, probably because it is trying to calculate the value for the whole CONUS-wide dataset rather than for the swept out area. wrf_to_kmz is meant to leave the data in its original projection until the kmz output, but somehow despite not being an all CONUS-wide dataset the value is still coming out as ~0.0. Also, trying out a WindNinja run on the wrf data, it fails to calculate the coordinateTransformationAngle, seems some more work might be needed for wrf wxModel datasets because of that lack of in-built projection information or something.