updated griddedInitialization to warp the input grids to the dem projection if the input grids are of a different projection, for the angleFromNorth stuff
that being said, the method works great for when DISABLE_ANGLE_FROM_NORTH_CALCULATION is set to TRUE, it warps the input grids to the dem coordinates without doing an angleFromNorth calculation, properly checking that the grids overlap the dem in the dem projection, running to completion treating the angles as geographic. But the method fails when DISABLE_ANGLE_FROM_NORTH_CALCULATION is set to FALSE, it is failing during the coordinateTransformationAngle calculation during GDALGetCenter(), because GDALGetGeoTransform() fails. I believe it is because the ascii grids are read in as gdal datasets rather than as ascii grids, the driver is probably not allowing use of a adfGeoTransform.

for issue #544