caught an error causing solpos solarIntensity calculations to fail for Solar/diurnal calculations, that was still letting the code continue, but resulted in the Solar/diurnal calculations using a single average flat terrain solarIntensity value, rather than the true per slope and aspect gridded solarIntensity values.
the initial flat terrain slope and aspect 0.0 value case for the Shade calculation was running properly with proper time variables, but the subsequent true slope and aspect gridded cellDiurnal calculations were being skipped because the time variables were dropped. solpos was throwing an error but the error was being ignored, and so the values were left as the single initial flat terrain slope and aspect 0.0 value.

this has apparently been going on for who knows how long, probably since the beginning, but thankfully comparisons of before and after results show that the diurnal calculations were definitely different, but not too terribly different, should just see some slight improvement to the results with this fix.