GUI weather station files with 24:00 midnight seg fault fix, by dropping an intermediate string to QDateTime() conversion and using the already in code boost methods. For issue #663
turns out the rest of the pointInitialization code was fine, because it would convert the 24:00 to 00:00 with day +1, so just needed to drop the intermediate qt stuff. Using a 25:00 time becomes 01:00 with day +1, so I believe that this fix should be robust. I did end up keeping some of the input time format checks though.