finally figured out how to get the progress bar 100% to properly display as 100%, as a full bar with 100% instead of as an empty bar with no number, including for the final set of messages and for the cancelled solve message. For issue #654
turned out to be as simple as "progressDialog->setAutoReset(false);". Apparently upon reaching 100%, the progress bar would restart/reset, where the restart value is pre- 0%, leading to an empty white bar with no % finished message. This forces the progress bar to stay at the final value instead of resetting the progress bar.