fix to 2nd QFutureWatcher progress bar attempt, for issue #654
needed to use a QMetaObject::invokeMethod() lambda function to be thread safe, when editing the QProgressDialog and other GUI widgets within QtConcurrent::run(). The past commit was technically working, but it would seg fault with a "Trace/breakpoint trap (core dumped)" error every once in a while

also got rid of the message duplication/overwriting problem that was causing me troubles from the past commit. Making the QProgressDialog and writeToConsole function calls thread safe seems to allow edited messages, from the QtConcurrent::run() function, to now properly propagate through to the rest of the code