final wrap-up work to the 2nd QFutureWatcher progress bar attempt, for issue #654
changed the QtConcurrent::run() and finished() functions, from being lambda functions, to being regularly called functions, as is done in the surfaceinput.h/.cpp and pointinitializationinput.h/.cpp qt6 gui code. In the process, learned that "this" can actually be passed in to QtConcurrent::run() functions, but only if the function to be ran is of non-static type, static seems to chop off or lose the member function-ness of the code.

Normally, it is very much not desired to pass in "this" to a QtConcurrent::run() function, but in this case, the function needed access to QProgressDialog and writeToConsole() GUI functions, within the QtConcurrent::run() worker threads. Passing in "this" also led to interesting access of other useful MainWindow stuff, like "ui" and MainWindow.h variables like maxProgress though.

the QMetaObject::invokeMethod() instances within QtConcurrent::run() are still lambda functions, but I had to use them to be able to safely run calls to QProgressDialog and writeToConsole() GUI functions, QMetaObject::invokeMethod() appears to be the method to make such GUI calls thread safe.