2nd attempt at upgrading ninjaCom methods, to use a callback function instead of QSocketNotifier or emit SIGNALS. For issue #666
note that this method is basically the same as using a GDALProgressFunction(), like is done for the wxModel progress bar, just I guess we are defining the methods ourselves. I disabled the use of the QSocketNotifier FILE* messages, and the stdout log printing in ninjaCom, for the testing of this method.