1st attempt at making the GUI ninjaError communications be done directly by ninjaCom rather than via thrown error messages, for issue #639, related to issue #666
The first step was to reverse the changes to windninja.cpp from commit 5e73e5cd7c512f8ba2db4f701c5583d990b14a6f, where the error messages were passed to the GUI through the C-API. The second step was to edit ninjaArmy.cpp to use ninjaCom instead of std::cout for each instance of "throw". The third step was to do a bit of cleanup to mainWindow.cpp, to avoid duplicating error messages, which were overwriting the ninjaCom changes to the progressDialog. I also did a bit of cleanup to ninjaCom to make messages a bit cleaner, some of which was stuff that I accidentally dropped from the previous ninjaCom code.

when editing ninjaArmy.cpp, it wasn't always clear which ninjaIdx to use for a given ninjaCom message, I tried to always start out using "ninjas[ninjas.size()-1]->ninjaCom()", adjusting to use the corresponding ninjaIdx where needed in a given section of code. The "i" was usually within scope for sections that required "ninjas[i]->ninjaCom()" but not so for all sections, there were also a few places where the "ninjas.size()" wasn't even defined yet. But I'm pretty sure that I got the ninjaIdx correct for the sections that I edited, more consideration should be made for non-solver sections where "return" is used instead of "throw".

Note that for instances of "throw" contained in a try/catch statement, a ninjaCom message wasn't required in the "try" part, because it gets taken care of in the "catch" part, don't want to duplicate messages.