took ascii2png() colorRamp stuff, from hard coded values, to CPLConfigOptions. This is very usable for smokejumper simulations now.
That being said, there is an issue with the units, ascii_grid.h cannot include ninjaUnits.h to allow unit conversion of the default "colorRampType" of "specificVals" color break values, it would cause a circular dependency. But, the workaround is to just manually update the color break values with the corresponding unit converted values using CPLConfigOptions.